Rename pet_surface_longitudinal.py to pet_surface_longitudinal_cli.py
